requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.8
argon2-cffi==21.1.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
Bottleneck==1.3.2
bunkai==1.5.2
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
clang==5.0
click==8.0.1
colorlog==6.4.1
cycler==0.10.0
dataclasses-json==0.6.3
dataframe-image==0.1.1
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
demoji==1.1.0
dill==0.3.4
emoji==1.4.2
emojis==0.7.0
entrypoints==0.3
fasttext==0.9.2
filelock==3.0.12
flatbuffers==1.12
fsspec==2021.8.1
fugashi==1.1.1
future==0.18.2
gast==0.4.0
gensim==4.1.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.40.0
h5py==3.1.0
htmlmin==0.1.12
huggingface-hub==0.0.16
hydra-colorlog==1.1.0
hydra-core==1.1.1
idna==3.2
ImageHash==4.2.1
install==1.3.4
ipadic==1.0.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
Janome==0.5.0
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.0.2
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
Markdown==3.3.4
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mecab==0.996.3
mecab-python3==1.0.4
missingno==0.5.0
mistune==0.8.4
mojimoji==0.0.11
more-itertools==10.1.0
multidict==5.1.0
multimethod==1.4
mypy-extensions==1.0.0
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
neologdn==0.5.1
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
omegaconf==2.1.1
opt-einsum==3.3.0
oseti==0.4.2
packaging==21.0
pandarallel==1.5.2
pandas==1.3.2
pandas-profiling==3.0.0
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==8.3.2
plac==1.3.3
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.7.1
pycparser==2.20
pydantic==1.8.2
pyDeprecate==0.3.1
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytorch-lightning==1.4.6
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
regex==2021.8.28
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
sacremoses==0.0.45
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
sengiri==0.2.1
sentencepiece==0.1.96
six==1.15.0
sklearn==0.0
smart-open==5.2.1
soupsieve==2.2.1
Spans==1.1.1
tangled-up-in-unicode==0.1.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==2.2.0
tokenizers==0.10.3
toml==0.10.2
torch==1.9.0
torch-tb-profiler==0.2.1
torchaudio==0.9.0
torchmetrics==0.5.1
torchtext==0.10.0
torchvision==0.10.0
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
transformers==4.10.2
typing-extensions==3.7.4.3
typing-inspect==0.9.0
unidic==1.0.3
unidic-lite==1.0.8
urllib3==1.26.6
visions==0.7.1
wasabi==0.8.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
yarl==1.6.3
